pyproject.toml
[tool.poetry]
name = "aws_serverless_cron"
version = "1.0.0"
description = "This repository provides a working, deployable, open source based, Serverless cron jobs"
authors = ["Ran Isenberg"]
classifiers=[
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.10",
]
repository="https://github.com/ran-isenberg/aws-lambda-handler-cookbook"
readme = "README.md"
keywords = ["aws lambda cron job", "AWS eventbridge scheduler", "AWS eventbridge rules", "serverless cron job"]
license = "MIT-0"

[tool.poetry.dependencies]
python = "^3.10.0"
pydantic = {version = "^2.0.0"}
email-validator = {version = "*"}
aws-lambda-powertools =  {git = "https://github.com/aws-powertools/powertools-lambda-python", rev = "poc/pydanticv2", extras = ["tracer"]}
cachetools = "*"
boto3 = "^1.26.125"
tenacity = "^8.2.2"

[tool.poetry.dev-dependencies]
# CDK
service-cdk = {path = "cdk", develop = true}
aws-cdk-lib = ">=2.75.0"
constructs = ">=10.0.0"
cdk-nag = ">2.0.0"
"aws-cdk.aws-lambda-python-alpha" = "^2.75.0-alpha.0"
# DEV
pytest = "*"
pytest-mock = "*"
pycodestyle = "*"
pytest-cov = "*"
pytest-html = "*"
python-dateutil = "*"
python-dotenv = "*"
GitPython = "*"
yapf = "*"
radon = "*"
xenon = "*"
pre-commit = "*"
flake8 = "*"
isort = "*"
mkdocs-material = "*"
mkdocs-git-revision-date-plugin = "*"
types-cachetools = "*"
mypy = "*"
types-requests = "*"
toml = "*"
